Switch to italic headers
Small caps feel more pretentious, and also don't work well when the
title contains numerals: the numerals are taller than the small-caps
letters. If I could reliably switch to old-style numerals just in the
headings, that would be better, but that's hard to do here without using
fontspec and OpenType fonts.